+++ cfrate.egg-info/PKG-INFO
@@ -1,0 +1,29 @@
+Metadata-Version: 2.1
+Name: cfrate
+Version: 0.0.5
+Summary: play a file sound - song- once the rating changes at all the participants  of some round. Also if a handle was given the program will output the the changes to that handle
+Home-page: https://github.com/theunderdog/Codeforces-rateDetector
+Author: theunderdog
+Author-email: ahmedbonumstelio@gmail.com
+License: UNKNOWN
+Description: # Codeforces-rateDetector
+        play a file sound - song- once the rating changes at all the participants  of some round. Also if a handle was given the program will output the the changes to that handle
+        <h4>This is only working on windows; However line 35 can be tweaked to work on linux - I just don't know how Â¯\_O_/Â¯ -</h4>
+        <h3>How to use:<h3/>
+        <ol>
+        <li>navigate to the folder that contain this script</li>
+        <li>add the file to be played</li>
+         <li>name the sound file "so.mp3" or change the variable "file" in the script to the file name</li>
+         <li>(optional) Change the variables "secs" to the number of seconds between requests</li>
+         <li>run in the command line <code>python run.py [contest id] [handle(optional)]</code>
+         </ol>
+         <h4>example</h4>
+         <code>python run.py 1047 theunderdog</code>
+         <br>
+         <code>python run.py 1047</code>
+        
+Platform: UNKNOWN
+Classifier: Programming Language :: Python :: 3
+Classifier: License :: OSI Approved :: MIT License
+Classifier: Operating System :: OS Independent
+Description-Content-Type: text/markdown
