+++ cfrate.egg-info/SOURCES.txt
@@ -1,0 +1,9 @@
+README.md
+setup.py
+cfrate/__init__.py
+cfrate/command.py
+cfrate.egg-info/PKG-INFO
+cfrate.egg-info/SOURCES.txt
+cfrate.egg-info/dependency_links.txt
+cfrate.egg-info/entry_points.txt
+cfrate.egg-info/top_level.txt
