+++ cfrate.egg-info/entry_points.txt
@@ -1,0 +1,3 @@
+[console_scripts]
+cfrate = cfrate.command:main
+
