+++ cfrate/command.py
@@ -1,0 +1,3 @@
+import cfrate
+def main():
+    cfrate.main()
